More on arithmetic functions
* All functions from The AWK Book table 2-6 are supported (except rand/srand)
* Floating number literals are now accepted by parser

+++ Lang/Hawk/AST.hs
@@ -2,7 +2,7 @@ module Lang.Hawk.AST where
 
 type Op = String
 
-data Literal = LitNumeric Integer
+data Literal = LitNumeric Double
              | LitStr String
              | LitRE String
                deriving (Eq, Show)
